# Optimize constructors of ImmutableTreeSet.