--- a/eclipse-collections/src/main/java/org/eclipse/collections/impl/set/sorted/immutable/ImmutableTreeSet.java
+++ b/eclipse-collections/src/main/java/org/eclipse/collections/impl/set/sorted/immutable/ImmutableTreeSet.java
@@ -15,6 +15,7 @@ import java.util.Arrays;
 import java.util.Comparator;
 import java.util.ConcurrentModificationException;
 import java.util.Iterator;
+import java.util.Objects;
 import java.util.Set;
 import java.util.SortedSet;
 import java.util.concurrent.ExecutorService;
@@ -33,7 +34,6 @@ import org.eclipse.collections.api.set.sorted.ImmutableSortedSet;
 import org.eclipse.collections.api.set.sorted.MutableSortedSet;
 import org.eclipse.collections.api.set.sorted.ParallelSortedSetIterable;
 import org.eclipse.collections.api.set.sorted.SortedSetIterable;
-import org.eclipse.collections.impl.block.factory.Comparators;
 import org.eclipse.collections.impl.factory.SortedSets;
 import org.eclipse.collections.impl.lazy.AbstractLazyIterable;
 import org.eclipse.collections.impl.lazy.parallel.AbstractBatch;
@@ -47,7 +47,6 @@ import org.eclipse.collections.impl.lazy.parallel.set.sorted.SelectSortedSetBatc
 import org.eclipse.collections.impl.lazy.parallel.set.sorted.SortedSetBatch;
 import org.eclipse.collections.impl.list.mutable.FastList;
 import org.eclipse.collections.impl.map.mutable.ConcurrentHashMap;
-import org.eclipse.collections.impl.set.sorted.mutable.TreeSortedSet;
 import org.eclipse.collections.impl.utility.ArrayIterate;
 import org.eclipse.collections.impl.utility.ListIterate;
 import org.eclipse.collections.impl.utility.internal.InternalArrayIterate;
@@ -62,42 +61,75 @@ final class ImmutableTreeSet<T>
     private final T[] delegate;
     private final Comparator<? super T> comparator;
 
-    private ImmutableTreeSet(SortedSet<T> sortedSet)
+    private ImmutableTreeSet(T[] input, Comparator<? super T> inputComparator, boolean isSortedAndUnique)
     {
-        this.delegate = (T[]) sortedSet.toArray();
-        this.comparator = sortedSet.comparator();
-    }
-
-    private ImmutableTreeSet(T[] delegate, Comparator<? super T> comparator)
-    {
-        if (comparator == null)
+        if (ArrayIterate.contains(input, null))
         {
-            comparator = Comparators.naturalOrder();
+            throw new NullPointerException("Input array contains nulls!");
         }
-        for (int i = delegate.length - 1; i > 0; i--)
+
+        if (isSortedAndUnique)
         {
-            if (comparator.compare(delegate[i - 1], delegate[i]) >= 0)
+            for (int i = input.length - 1; i > 0; i--)
             {
-                throw new ConcurrentModificationException("Input Array expected to be sorted, but was not!");
+                int compare = inputComparator == null
+                        ? ((Comparable<? super T>) input[i - 1]).compareTo(input[i])
+                        : inputComparator.compare(input[i - 1], input[i]);
+                if (compare >= 0)
+                {
+                    throw new ConcurrentModificationException("Input Array expected to be sorted, but was not!");
+                }
             }
         }
-        this.delegate = delegate;
-        this.comparator = comparator;
+        else
+        {
+            if (input.length > 0)
+            {
+                Arrays.sort(input, inputComparator);
+                T[] unique = (T[]) new Object[input.length];
+                unique[0] = input[0];
+
+                if (inputComparator == null && !(input[0] instanceof Comparable))
+                {
+                    throw new ClassCastException("Comparator is null and input does not implement Comparable!");
+                }
+
+                int uniqueCount = 1;
+                for (int i = 1; i < input.length; i++)
+                {
+                    int compare = inputComparator == null
+                            ? ((Comparable<? super T>) unique[uniqueCount - 1]).compareTo(input[i])
+                            : inputComparator.compare(unique[uniqueCount - 1], input[i]);
+                    if (compare < 0)
+                    {
+                        unique[uniqueCount] = input[i];
+                        uniqueCount++;
+                    }
+                }
+                if (uniqueCount < input.length)
+                {
+                    input = Arrays.copyOf(unique, uniqueCount);
+                }
+            }
+        }
+
+        this.delegate = input;
+        this.comparator = inputComparator;
     }
 
     public static <T> ImmutableSortedSet<T> newSetWith(T... elements)
     {
-        return new ImmutableTreeSet<>(TreeSortedSet.newSetWith(elements));
+        return new ImmutableTreeSet<>(elements.clone(), null, false);
     }
 
     public static <T> ImmutableSortedSet<T> newSetWith(Comparator<? super T> comparator, T... elements)
     {
-        return new ImmutableTreeSet<>(TreeSortedSet.newSetWith(comparator, elements));
+        return new ImmutableTreeSet<>(elements.clone(), comparator, false);
     }
 
     public static <T> ImmutableSortedSet<T> newSet(SortedSet<T> set)
     {
-        return new ImmutableTreeSet<>((T[]) set.toArray(), set.comparator());
+        return new ImmutableTreeSet<>((T[]) set.toArray(), set.comparator(), true);
     }
 
     @Override